+++ NEWS.md
@@ -10,3 +10,7 @@
 
 * Added support for working directly from VCF files
 * Added SNV, Insertion & Deletion calling (`calvInDel`)
+
+# transmogR 0.1.3
+
+* Added UpSet plot for showing variants by gene/transcript